Add retry logic to 'BrowserManager.connectToPeer()'
Before this commit: if inviting a remote peer to connect fails,
connectToPeer() calls the completion handler right away even for
transient errors letting the application code to decide to trigger
a retry logic.
This commit changes that behavior, so if the error is a transient one,
connectToPeer() retries for a few times before bailing out and passing
a fatal error.
Other fixes/changes included in this commit:
 - reduce the invitePeerTimeout to 1 second (was 5)
 - properly release the Session.mcSession instance
 - properly release the BrowserRelays.nonTCPSession instance
 - return the 'Session disconnected' error
 - remove race condition if a peer is lost while attempting to connect
 - rename some variables
 - more debug log