Increase the 'invite to connect' timeout to 2.5s.
It seems that under load the MPCF is slow to establish connections.
The 1s timeout worked just fine when running individual tests, but when
running the full ThaliTest tests, time to time there were connections
failure without the MPCF handshaking even starting.
Increasing the timeout to 2.5 made all the test pass for several times
in a row.